TestNextAgreementRedirect: this really belongs in test_agreements.py to correspond to where its implementation is
