+++ tests/app/main/views/test_suppliers.py
@@ -1,5 +1,5 @@
 from six import BytesIO
-from six.moves.urllib.parse import urlparse, urlsplit, parse_qs
+from six.moves.urllib.parse import urlsplit
 import mock
 from freezegun import freeze_time
 from lxml import html
@@ -1377,138 +1377,3 @@ class TestCorrectButtonsAreShownDependingOnContext(LoggedInApplicationTest):
         assert len(document.xpath(
             "//a[@href='/admin/suppliers/1234/agreements/g-cloud-8/next'][normalize-space(string())='Next agreement']"
         )) == 1
-
-
-@mock.patch('app.main.views.agreements.data_api_client')
-class TestNextAgreementRedirect(LoggedInApplicationTest):
-    user_role = 'admin'
-
-    @property
-    def dummy_supplier_frameworks(self):
-        # a property so we ensure we get a new clone every time we request it
-        return {
-            "supplierFrameworks": [
-                {
-                    "supplierId": 4321,
-                    "frameworkSlug": "g-cloud-8",
-                    "agreementStatus": "signed",
-                },
-                {
-                    "supplierId": 1234,
-                    "frameworkSlug": "g-cloud-8",
-                    "agreementStatus": "on-hold",
-                },
-                {
-                    "supplierId": 31415,
-                    "frameworkSlug": "g-cloud-8",
-                    "agreementStatus": "signed",
-                },
-                {
-                    "supplierId": 27,
-                    "frameworkSlug": "g-cloud-8",
-                    "agreementStatus": "signed",
-                },
-                {
-                    "supplierId": 141,
-                    "frameworkSlug": "g-cloud-8",
-                    "agreementStatus": "countersigned",
-                },
-                {
-                    "supplierId": 151,
-                    "frameworkSlug": "g-cloud-8",
-                    "agreementStatus": "on-hold",
-                },
-                {
-                    "supplierId": 99,
-                    "frameworkSlug": "g-cloud-8",
-                    "agreementStatus": "approved",
-                },
-            ],
-        }
-
-    def test_happy_path(self, data_api_client):
-        data_api_client.find_framework_suppliers.return_value = self.dummy_supplier_frameworks
-        res = self.client.get('/admin/suppliers/1234/agreements/g-cloud-8/next')
-        assert res.status_code == 302
-
-        parsed_location = urlparse(res.location)
-        assert parsed_location.path == "/admin/suppliers/31415/agreements/g-cloud-8"
-        assert parse_qs(parsed_location.query) == {}
-
-    def test_unknown_supplier_returns_404(self, data_api_client):
-        data_api_client.find_framework_suppliers.return_value = self.dummy_supplier_frameworks
-        res = self.client.get('/admin/suppliers/999/agreements/g-cloud-8/next')
-        assert res.status_code == 404
-
-    def test_final_supplier_redirects_to_list(self, data_api_client):
-        data_api_client.find_framework_suppliers.return_value = self.dummy_supplier_frameworks
-        res = self.client.get('/admin/suppliers/99/agreements/g-cloud-8/next')
-        assert res.status_code == 302
-
-        parsed_location = urlparse(res.location)
-        assert parsed_location.path == "/admin/agreements/g-cloud-8"
-        assert parse_qs(parsed_location.query) == {}
-
-    def test_status_filtered_simple_status(self, data_api_client):
-        data_api_client.find_framework_suppliers.return_value = self.dummy_supplier_frameworks
-        res = self.client.get('/admin/suppliers/1234/agreements/g-cloud-8/next?status=on-hold')
-        assert res.status_code == 302
-
-        parsed_location = urlparse(res.location)
-        assert parsed_location.path == "/admin/suppliers/151/agreements/g-cloud-8"
-        assert parse_qs(parsed_location.query) == {"next_status": ["on-hold"]}
-
-        # if there weren't any non-status-filtering find_framework_suppliers calls, the view won't have been able to
-        # get it right
-        assert any(
-            (not ca_kwargs.get("status"))
-            for ca_args, ca_kwargs in data_api_client.find_framework_suppliers.call_args_list
-        )
-
-    def test_status_filtered_compound_status(self, data_api_client):
-        data_api_client.find_framework_suppliers.return_value = self.dummy_supplier_frameworks
-        res = self.client.get('/admin/suppliers/141/agreements/g-cloud-8/next?status=approved,countersigned')
-        assert res.status_code == 302
-
-        parsed_location = urlparse(res.location)
-        assert parsed_location.path == "/admin/suppliers/99/agreements/g-cloud-8"
-        assert parse_qs(parsed_location.query) == {"next_status": ["approved,countersigned"]}
-
-        # if there weren't any non-status-filtering find_framework_suppliers calls, the view won't have been able to
-        # get it right
-        assert any(
-            (not ca_kwargs.get("status"))
-            for ca_args, ca_kwargs in data_api_client.find_framework_suppliers.call_args_list
-        )
-
-    def test_status_filtered_different_status_supplier(self, data_api_client):
-        data_api_client.find_framework_suppliers.return_value = self.dummy_supplier_frameworks
-        res = self.client.get('/admin/suppliers/1234/agreements/g-cloud-8/next?status=signed')
-        assert res.status_code == 302
-
-        parsed_location = urlparse(res.location)
-        assert parsed_location.path == "/admin/suppliers/31415/agreements/g-cloud-8"
-        assert parse_qs(parsed_location.query) == {"next_status": ["signed"]}
-
-        # if there weren't any non-status-filtering find_framework_suppliers calls, the view won't have been able to
-        # get it right
-        assert any(
-            (not ca_kwargs.get("status"))
-            for ca_args, ca_kwargs in data_api_client.find_framework_suppliers.call_args_list
-        )
-
-    def test_status_filtered_final_with_status_redirects_to_list(self, data_api_client):
-        data_api_client.find_framework_suppliers.return_value = self.dummy_supplier_frameworks
-        res = self.client.get('/admin/suppliers/151/agreements/g-cloud-8/next?status=on-hold')
-        assert res.status_code == 302
-
-        parsed_location = urlparse(res.location)
-        assert parsed_location.path == "/admin/agreements/g-cloud-8"
-        assert parse_qs(parsed_location.query) == {"status": ["on-hold"]}
-
-        # if there weren't any non-status-filtering find_framework_suppliers calls, the view won't have been able to
-        # get it right
-        assert any(
-            (not ca_kwargs.get("status"))
-            for ca_args, ca_kwargs in data_api_client.find_framework_suppliers.call_args_list
-        )
